LWE/requirements.txt
openai>=1.40.0
google-genai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
PyYAML>=6.0
torch>=2.0.0
tqdm>=4.65.0
datasets>=2.14.0
Pillow>=10.0.0
